+++ src/main/java/com/example/catchtable/domain/PickupTime.java
@@ -11,12 +11,12 @@ import java.sql.Time;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PickupTime {
 
-  private Integer id;
+  private Long id;
   private Integer restaurantId;
   private Time time;
 
   // 정적 팩토리 메소드
-  public static PickupTime fromEntity(Integer id, Time time, Integer restaurantId) {
+  public static PickupTime fromEntity(Long id, Time time, Integer restaurantId) {
     PickupTime pickupTime = new PickupTime();
     pickupTime.id = id;
     pickupTime.time = time;
