+++ src/main/java/com/example/catchtable/domain/ReservationStatus.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ReservationStatus {
 
-  private Integer id;
+  private Long id;
   private String type;
 
   // 정적 팩토리 메소드
-  public static ReservationStatus fromEntity(final Integer id, final String type) {
+  public static ReservationStatus fromEntity(final Long id, final String type) {
     ReservationStatus reservationStatus = new ReservationStatus();
     reservationStatus.id = id;
     reservationStatus.type = type;
