+++ src/main/java/com/example/catchtable/repository/PaymentHistoryRepository.java
@@ -28,7 +28,7 @@ public class PaymentHistoryRepository {
 
   private final RowMapper<PaymentHistory> paymentHistoryRowMapper = (rs, rowNum) ->
       PaymentHistory.fromEntity(
-          rs.getInt("id"),
+          rs.getLong("id"),
           rs.getInt("method"),
           rs.getInt("amount"),
           rs.getInt("status"),
@@ -57,7 +57,7 @@ public class PaymentHistoryRepository {
       return ps;
     }, keyHolder);
     Number key = keyHolder.getKey();
-    return findById(Objects.requireNonNull(key).intValue());
+    return findById((long) Objects.requireNonNull(key).intValue());
   }
 
   private Optional<PaymentHistory> update(PaymentHistory entity) {
@@ -72,13 +72,13 @@ public class PaymentHistoryRepository {
     return findAll(entities);
   }
 
-  public Optional<PaymentHistory> findById(Integer id) {
+  public Optional<PaymentHistory> findById(Long id) {
     String sql = "SELECT * FROM payment_history WHERE id = ?";
     List<PaymentHistory> result = jdbcTemplate.query(sql, paymentHistoryRowMapper, id);
     return result.isEmpty() ? Optional.empty() : Optional.of(result.get(0));
   }
 
-  public boolean existsById(Integer id) {
+  public boolean existsById(Long id) {
     String sql = "SELECT count(*) FROM payment_history WHERE id = ?";
     var result = jdbcTemplate.queryForObject(sql, Long.class, id);
     return Optional.ofNullable(result).orElse(0L) > 0;
@@ -105,7 +105,7 @@ public class PaymentHistoryRepository {
     return Optional.ofNullable(result).orElse(0L);
   }
 
-  public void deleteById(Integer id) {
+  public void deleteById(Long id) {
     String sql = "DELETE FROM payment_history WHERE id = ?";
     jdbcTemplate.update(sql, id);
   }
