+++ src/main/java/com/example/catchtable/repository/PaymentRepository.java
@@ -28,7 +28,7 @@ public class PaymentRepository {
 
   private final RowMapper<Payment> paymentRowMapper = (rs, rowNum) ->
       Payment.fromEntity(
-          rs.getInt("id"),
+          rs.getLong("id"),
           rs.getInt("amount"),
           rs.getInt("order_id"),
           rs.getTimestamp("created_at"),
@@ -57,7 +57,7 @@ public class PaymentRepository {
       return ps;
     }, keyHolder);
     Number key = keyHolder.getKey();
-    return findById(Objects.requireNonNull(key).intValue());
+    return findById((long) Objects.requireNonNull(key).intValue());
   }
 
   private Optional<Payment> update(Payment entity) {
@@ -71,13 +71,13 @@ public class PaymentRepository {
     return findAll(entities);
   }
 
-  public Optional<Payment> findById(Integer id) {
+  public Optional<Payment> findById(Long id) {
     String sql = "SELECT * FROM payment WHERE id = ?";
     List<Payment> result = jdbcTemplate.query(sql, paymentRowMapper, id);
     return result.isEmpty() ? Optional.empty() : Optional.of(result.get(0));
   }
 
-  public boolean existsById(Integer id) {
+  public boolean existsById(Long id) {
     String sql = "SELECT count(*) FROM payment WHERE id = ?";
     var result = jdbcTemplate.queryForObject(sql, Long.class, id);
     return Optional.ofNullable(result).orElse(0L) > 0;
@@ -104,7 +104,7 @@ public class PaymentRepository {
     return Optional.ofNullable(result).orElse(0L);
   }
 
-  public void deleteById(Integer id) {
+  public void deleteById(Long id) {
     String sql = "DELETE FROM payment WHERE id = ?";
     jdbcTemplate.update(sql, id);
   }
